refactor: Refactor entity models and access rules
- Moved `NoActiveRevisionsError` import to `include.exceptions.misc` for better organization.
- Updated import paths for `Document`, `DocumentRevision`, and `Folder` to reflect new structure in `obj.py`.
- Created `base.py` for shared base functionality among entity models.
- Introduced `metadata.py` for document metadata handling.
- Added `obj.py` to define core entity models including `Document`, `DocumentRevision`, `Folder`, and their access rules.
- Implemented `batch_count_other_revisions` in `count.py` for centralized reference counting.
- Refactored `purge.py` to utilize new counting method.
- Adjusted `document.py` and `search.py` to align with new import paths.
